Introduce FaLayersCounterDirective and refactor FaLayersCounterComponent to use host directives
